Added logout logic to the button on the search bar and disabled connexion page if user is connected
