+++ src/pages/connexion.js
@@ -1,4 +1,4 @@
-import React, { useRef, useState } from "react";
+import React, { useRef, useState, useEffect } from "react";
 
 import { Link, useNavigate } from "react-router-dom";
 
@@ -12,10 +12,17 @@ import { useAuth } from "../contexts/AuthContext";
 const Connexion = () => {
   const emailRef = useRef();
   const passwordRef = useRef();
-  const { login, logout } = useAuth();
+  const { login, currentUser } = useAuth();
   const [error, setError] = useState("");
   const [loading, setLoading] = useState(false);
   const navigate = useNavigate();
+
+  // Redirect to home if user is already logged in
+  useEffect(() => {
+    if (currentUser) {
+      navigate("/");
+    }
+  }, [currentUser, navigate]);
 
   // login
   async function handleSubmit(e) {
@@ -29,22 +36,10 @@ const Connexion = () => {
     } catch (error) {
       // Show the error message from Firebase
       setError(error.message);
+      console.error(error.message);
     }
 
     setLoading(false);
-  }
-
-  // logout
-  async function handleLogout() {
-    setError("");
-
-    try {
-      await logout();
-      navigate("/connexion");
-      setError("Vous êtes déconnecté");
-    } catch (error) {
-      setError(error.message); // Show the error message from Firebase
-    }
   }
 
   return (
@@ -129,6 +124,7 @@ const Connexion = () => {
             <Button
               type="submit"
               className="flex w-full justify-center bg-indigo-500 hover:bg-indigo-400 focus-visible:outline-indigo-500 text-white"
+              disabled={loading}
             >
               Se connecter
             </Button>
@@ -158,18 +154,6 @@ const Connexion = () => {
             >
               <img className="h-5 w-5" src={google_icon} alt="Google" />
               Continuer avec Google
-            </Button>
-          </div>
-
-          {/* BUTTON LOGOUT */}
-          <div className="mt-6 grid grid-cols-1 gap-4">
-            <Button
-              as="a" // Specify that it should render as an anchor tag
-              href="#"
-              className="flex w-full items-center justify-center gap-3 bg-red-600 text-gray-900 focus-visible:outline-[#1D9BF0]"
-              onClick={handleLogout}
-            >
-              Logout
             </Button>
           </div>
         </div>
